Remove redundant else
An `else` at the end of a function or method, where the `if` just returned a value, can be removed in favour of just giving the final statement (here, another `return` statement).

An `else` that only contains a new `if` condition, can have the `if` merged with the parent `else`.