Home page: Check is_front_page() earlier
Because the front page is a special circumstance for Pages, then the `is_front_page()` check needs to come before the check for regular page types.

With the check now earlier, the headline and URL will correctly represent the home page of a site (site name, rather than Page title).

Fixes #107.

Also: Handle misconfigured front page settings

When the settings are to show a page on the front, but not Page has been set, then `is_front_page()` returns false.

In terms of generating the Parsely metadata though, whatever is showing at the root of the site, should still show the site name and home URL.