Simple JS WP build env + prettier & eslint
This is a step toward a more robust JS environment for the front end and admin (ES6+ syntax, importing npm packages, etc.).

It leverages the wp-scripts package so we don't need a custom webpack config (yet).

Additionally, this includes prettier and eslint configurations to enable IDE auto-formatters.

Right now, the `src/index.js` file is only a placeholder for the entrypoint. We'll want to replace it with an actual implementation prior to sourcing the built file in the plugin.