Widget: Allow no boost to scores
The [API Docs](https://www.parse.ly/help/api/recommendations#get-related) say that the default value for boost is undefined (ie. score only).

However, the dropdown field in the widget had no option to choose this; so there was always a boost value.

A "No boost" option has now been added, and the `boost` parameter is not included if it is set to this.

There is no change to the default value of this field - it is still `views` (Page views).

Existing sites would also not be affected, as their existing widget settings would continue to be used, regardless of what new options there now were.

Closes #161.