Add $post_id to wp_parsely_permalink filter hook
The `wp_parsely_permalink` filter only provided the existing permalink to be filtered, and a $post variable, which is really just a "post" or "nonpost" string.

We now provide the $post_id, which is the ID of the specific post that the get_current_url() is working with. If the `$post_id` was not provided to the parent function, then the value will be 0, and the permalink will be for the `global $post`.

Closes #168.