Widget: Hide form when settings are missing
The widget requires the "Parse.ly Site ID" (API key) and "Parse.ly API Secret" to be populated.

If either field were not populated, the widget form would still display the regular widget settings, even though there was zero chance of the widget working correctly.

This has now switched to checking for the settings, and giving a message to the user with a link to the settings page.

See #156.