Fix requires_recrawl check for settings
Currently, settings are set to either true or false for `requires_recrawl` - however we just do an `isset` check to display the warning.  That means the warning displays for every case even if it’s not supposed to.

I thought about doing just `empty()` but I feel like `isset and true` is more readable so I went in favour of that.