Translations: Fix concatenated i18n strings
The `WordPress.WP.I18n.NonSingularStringLiteralText` PHPCS sniff checks if the `$text_arg` from i18n functions contains a single string literal, which is required to properly extract the string.

This fixes the arguments that contained concatenated strings.

Strings literals can't be broken onto multiple lines, as while the display is fine (browsers normalize the line break into a space), the translatable string in the .pot file contains `\n` and a number of `\t` tab characters, which is undesirable for translation contributors to deal with.