CI: Use OS with MySQL 5.7
ubuntu-20.04 / ubuntu-latest includes MySQL 8, which has issues with older versions of PHP. See https://www.php.net/manual/en/mysqli.requirements.php for more info.

There may be a workaround, but since ubuntu-16.04 already has the range of PHP versions that are in the matrix, then that can be used for now.