Remove custom test constant
Existing tests pass with and without this constant, which means it's not doing anything (which is a good thing).

The only thing that changed is the cache busting string changed when WP_DEBUG was defined (not also checks it is truthy), so we can use an existing WP test constant for that, instead of defining our own.

Fixes incorrectly named static variable, adds documentation to the filter.

Fixes #256.