Row action: add aria-label to make link accessible
To have the same link text (such as "Parse.ly stats") pointing to different URLs, is a bad thing for accessibility. Links can be taken out of context, so they need to be more differentiated and unique.

We follow the same principle as what WordPress core does for the existing row action links, and that is to include an `aria-label` attribute which includes the title of the post. Visually, the link looks the same.