Tests: Add wp-test-utils
WP Test Utils is a PHPUnit cross-version compatibility layer for testing plugins and themes build for WordPress.

It has several useful features, but the features we'll use to start with, are:

- Cross-version compatibility with PHPUnit 5.7 - 9.x via the [PHPUnit Polyfills](https://packagist.org/packages/yoast/phpunit-polyfills) package.
  _Note: WordPress Core limit tests to running on PHPUnit 7.5 max. However, using these polyfills we can already start using the up-to-date PHPUnit 9.x syntax, even though the tests don't use PHPUnit 9 yet._
- A custom autoloader which will selectively autoload the WP copies of the PHPUnit 9 native MockBuilder files for PHPUnit < 9 when run on PHP 8 to get round the use of the new reserved keyword `match` as was used in older versions of these files. Without this, we can't use the MockBuilder in PHPUnit.

In the future, we can split out unit tests from WP (integration) tests, and WP Test Utils provides testcases that have BrainMonkey (mocking library for WordPress) support as well as pre-defining some WP-specific constants that our code might be using.

The version of wp-test-utils is pinned to a commit-ref that addresses a [particular issue](https://github.com/Yoast/wp-test-utils/issues/3) when using the standard WP-CLI tests scaffolding directory structure.

The `cs-lint.yml` change is to stop the action failing on the warning that comes with using a commit-ref.

The bootstrap file changes ensure the custom autoloader works correctly.