Tests: Extract and improve FBIA test
The previous test was flawed, in that it didn't define a `$registry` variable, and the only checked the embed code for two URLs, one of which was dependent on the test setup.

The new test is less concerned about the specifics of the embed code (generally hard-coded), but does check for the presence of the API key (dynamic), and that there is at least one script component in it.

After checking the actual integration, it was clear that the `return` in the main logic was redundant.