Composer: Use tagged version of wp-test-utils
In #333, we temporarily used a commit ref for the `yoast/wp-test-utils` package version to address a particular issue. This issue has now been fixed and is available in a tagged release, so we can now use that.

We had also had to remove the `--strict` flag from the `composer validate` command in a CI workflow to allow the commit ref to be used; this has now been changed back as well.