Parsely: Move hooks out of constructor
Class constructors should help set the class up (i.e. populate properties), but that's it.

Moving the adding of action and filter hooks to a `run()` method, and calling that after class instantiation, means the class can now be instantiated in tests without having potential side-effects - for instance, the upgrade code can be moved to a new method and tested.

Fixes #280.