Deploy on published release
As per https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows#release, the workflow will only fire when we publish a release on GitHub. The `released` type will only fire for stable releases,  and not pre-release releases. 

This allows tagging and release of alpha and beta versions, without it automatically pushing to WPORG SVN which then result in sites pulling those in via plugin auto-updates.