PR Template: Remove checkboxes
Checkboxes on PRs are often made more visible as a list of things to check (one or more can and should be marked off) - and shouldn't be treated as a radio button (only one item selected from a list).

In this case, the type of fix being asked for, was in relation to Semantic Versioning, but it wasn't comprehensive enough (many PRs are Type: Maintenance), and so labels would be a better choice for this by the repo owners, rather than just contributors.