Settings: Avoid blocking invalid form submissions
While we want to programattically indicate that a field is required, we don't want different browser implementations of client-side validation to block the form submission. It leads to inconsistent behaviour across browsers, isn't overly accessible to assistive technologies, and we already have a warning message that the single required field is required.