Add Screen Options to toggle settings sections
The settings page can be long so by using the Screen Options feature we can enable users to hide those settings that either require a recrawl, or advanced settings that would wipe the data.

Note that for the $_POST data to be captured, the Settings_Page class is instantiated on the `init` hook instead of the (internal-only) `_admin_menu`. The `run()` method of the Settings_Page class hooks into specific admin-page related hooks, so there is no substantial impact to front-end requests.