Settings page: Show error message when duplicate tracking selected
If the user selects the same element in the "Post Types to Track" and "Page Types to Track" sections, an error message will be shown to them telling them that the element should be selected only in one of the sections.

The reasoning behind this is that we should not allow duplicate tracking.