Composer: permit composer plugins
The `dealerdirect/phpcodesniffer-composer-installer` and `composer/installers` Composer plugins are used to register an external PHPCS standards with PHPCS, and allow the plugin to be installed in a `wp-content/plugins` directory.

As of Composer 2.2.0, Composer plugins need to be explicitly allowed to run. This adds the necessary configuration for that.

Any globally-installed Composer plugin should be configured in the global `composer.json`.

`composer normalize` also run on the file, which fixes a non-alphabetical script.

Also refresh the list of script descriptions.

Refs:

https://blog.packagist.com/composer-2-2/#more-secure-plugin-execution