Fix failing E2E tests
While performing E2E testing, the function changeKeysState() would result in a WordPress critical error which resulted in tests timing out. This would only happen on the testing environment as this wasn't reproducible on the development environment.

It would seem that under the test environment, some arrays weren't being set and this was making code inside the validate_options_post_type_tracking() function to crash, resulting in the WordPress critical error. I was never able to reproduce this in my development environment, but the fix is simple and allows everything to work correctly in all environments.