Remove instantiations from inside Related_API_Proxy
Due to previous changes, we can now inject a proxy object into the constructor of Related_API_Proxy, and don't have to instantiate the proxy, cache, and cached_proxy objects inside this Related_API_Proxy class. This means that the Related_API_Proxy class no longer has knowledge that a cache behaviour even exists for the proxy class, let alone where the result might be cached.

All the REST API class knows, is that it is given a proxy class, and it defers to that when items need to be retrieved.

The decision to use a cache, and what cache implementation to use, is now handled only in the composable root / main plugin file.

The integration test passes in a proxy object without the caching behaviour.