Switch messy integration test to unit test
Instead of stubbing out WPGraphQL functions under a `Parsely\Endpoints` namespace, and having it set public properties in the test class, let's use BrainMonkey to simply patch the missing external functions instead.

This commit adds a new unit test that includes a test for whether these functions defined in the WPGraphQL plugin are called. We use `Brain\Monkey\Functions\expect()` to name the functions we're expecting to be called, and then also state how many times we expect them to be called, and also what arguments (values or type) they should be called with. If in the course of running the test, anything differs, the test fails. Otherwise, it passes.

The `Brain\Monkey\Functions\when()` call is there to mock the WordPress-defined `__()` function, since in a unit test, WordPress isn't loaded. In this case, we don't care what it was called with, but we patch it to simple return the first argument that was passed to it (the string to be translated).