Adding GraphQL support (#710)
* Adding GraphQL support

* Update src/Endpoints/class-graphql-metadata.php

Co-authored-by: Gary Jones <gary.jones@automattic.com>

* Update src/Endpoints/class-graphql-metadata.php

Co-authored-by: Gary Jones <gary.jones@automattic.com>

* Moving GraphQL class check to wp-parsely.php

* Translate untranslated string

* Extracting resolve into function

* Extracting configuration into variable

* Update src/Endpoints/class-graphql-metadata.php

Co-authored-by: Chris Zarate <chris@zarate.org>

* Update src/Endpoints/class-graphql-metadata.php

Co-authored-by: Chris Zarate <chris@zarate.org>

* Update src/Endpoints/class-graphql-metadata.php

Co-authored-by: Chris Zarate <chris@zarate.org>

* Removing redundant line break

* Moving to rendered-only fields

* Returning tracker URL on graphql

* Hook to an interface

* Fixing tests

* Fix isTracked check

* Allowing removeWrappingTag

* Fixing tests

* Move directive to field resolver

1. Move `removeWrappingTag` directive to `jsonLd` field resolver.
2. Move the code that loads the data for both `jsonLd` and
   `repeatedMetas` to field resolvers so that they are only called when
   queried.
3. Return `null` instead of empty array if the post cannot be found,
   which will prevent subfields from being resolved. (This will never
   actually happen in practice.)

* Fix alignment

* Adding URL generation tests

* Adding GraphQL tests

* Mocking GraphQL functions

* Switch messy integration test to unit test

Instead of stubbing out WPGraphQL functions under a `Parsely\Endpoints` namespace, and having it set public properties in the test class, let's use BrainMonkey to simply patch the missing external functions instead.

This commit adds a new unit test that includes a test for whether these functions defined in the WPGraphQL plugin are called. We use `Brain\Monkey\Functions\expect()` to name the functions we're expecting to be called, and then also state how many times we expect them to be called, and also what arguments (values or type) they should be called with. If in the course of running the test, anything differs, the test fails. Otherwise, it passes.

The `Brain\Monkey\Functions\when()` call is there to mock the WordPress-defined `__()` function, since in a unit test, WordPress isn't loaded. In this case, we don't care what it was called with, but we patch it to simple return the first argument that was passed to it (the string to be translated).

* Apply suggestions from code review

Co-authored-by: Alex Cicovic <23142906+acicovic@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Alex Cicovic <23142906+acicovic@users.noreply.github.com>

* Updating comment

Co-authored-by: Gary Jones <gary.jones@automattic.com>
Co-authored-by: Chris Zarate <chris@zarate.org>
Co-authored-by: chriszarate <chris.zarate@automattic.com>
Co-authored-by: Alex Cicovic <23142906+acicovic@users.noreply.github.com>