Settings page: Remove media months query (#827)
The settings page calls `wp_enqueue_media()`, and this always runs a query to capture the month-year combinations that attachments were uploaded. This is only potentially useful when the logo is being selected, and even then, it's more likely that the user would upload a new logo or choose a recently uploaded one, than use the month-year filter to try and find an older upload (search by file is also available anyway).

The filter hook to disable this query has a note that this query can be expensive for large media libraries, so let's disable this query by default. Since we control the page, and won't be calling `wp_enqueue_media()` a second time, there's no benefit to re-enabling the month-year query after `wp_enqueue_media()` has been called.

Co-authored-by: Pau Argelaguet <pauarge@users.noreply.github.com>
Co-authored-by: Alex Cicovic <23142906+acicovic@users.noreply.github.com>