Rector: Add rector
Add Rector (https://getrector.org/) as a tool to help catch missed opportunities to improve code, particularly when upgrading to newer minimum PHP versions, or choosing to refactor to or from particularly coding styles.

The default configuration here is to apply all of the rules for the set up to PHP 7.1. It skips the long array to short array rule, since that conflicts with the WordPress Coding Style which the plugin follows.